+++ lib/service_actor/context.rb
@@ -1,31 +1,33 @@
-module Interactor
-  # Public: The object for tracking state of an Interactor's invocation. The
-  # context is used to initialize the interactor with the information required
-  # for invocation. The interactor manipulates the context to produce the result
+# frozen_string_literal: true
+
+module ServiceActor
+  # Public: The object for tracking state of a ServiceActor's invocation. The
+  # context is used to initialize the actor with the information required
+  # for invocation. The actor manipulates the context to produce the result
   # of invocation.
   #
   # The context is the mechanism by which success and failure are determined and
-  # the context is responsible for tracking individual interactor invocations
+  # the context is responsible for tracking individual actor invocations
   # for the purpose of rollback.
   #
   # The context may be manipulated using arbitrary getter and setter methods.
   #
   # Examples
   #
-  #   context = Interactor::Context.new
-  #   # => #<Interactor::Context>
+  #   context = ServiceActor::Context.new
+  #   # => #<ServiceActor::Context>
   #   context.foo = "bar"
   #   # => "bar"
   #   context
-  #   # => #<Interactor::Context foo="bar">
+  #   # => #<ServiceActor::Context foo="bar">
   #   context.hello = "world"
   #   # => "world"
   #   context
-  #   # => #<Interactor::Context foo="bar" hello="world">
+  #   # => #<ServiceActor::Context foo="bar" hello="world">
   #   context.foo = "baz"
   #   # => "baz"
   #   context
-  #   # => #<Interactor::Context foo="baz" hello="world">
+  #   # => #<ServiceActor::Context foo="baz" hello="world">
   class Context
     # Internal: Initialize a new Context with the given attributes.
     #
@@ -111,46 +113,47 @@ module Interactor
       @table
     end
     private :modifiable
-    # Internal: Initialize an Interactor::Context or preserve an existing one.
-    # If the argument given is an Interactor::Context, the argument is returned.
-    # Otherwise, a new Interactor::Context is initialized from the provided
+
+    # Internal: Initialize a ServiceActor::Context or preserve an existing one.
+    # If the argument given is a ServiceActor::Context, the argument is returned.
+    # Otherwise, a new ServiceActor::Context is initialized from the provided
     # hash.
     #
-    # The "build" method is used during interactor initialization.
+    # The "build" method is used during actor initialization.
     #
     # context - A Hash whose key/value pairs are used in initializing a new
-    #           Interactor::Context object. If an existing Interactor::Context
+    #           ServiceActor::Context object. If an existing ServiceActor::Context
     #           is given, it is simply returned. (default: {})
     #
     # Examples
     #
-    #   context = Interactor::Context.build(foo: "bar")
-    #   # => #<Interactor::Context foo="bar">
+    #   context = ServiceActor::Context.build(foo: "bar")
+    #   # => #<ServiceActor::Context foo="bar">
     #   context.object_id
     #   # => 2170969340
-    #   context = Interactor::Context.build(context)
-    #   # => #<Interactor::Context foo="bar">
+    #   context = ServiceActor::Context.build(context)
+    #   # => #<ServiceActor::Context foo="bar">
     #   context.object_id
     #   # => 2170969340
     #
-    # Returns the Interactor::Context.
+    # Returns the ServiceActor::Context.
     def self.build(context = {})
       self === context ? context : new(context)
     end
 
-    # Public: Whether the Interactor::Context is successful. By default, a new
+    # Public: Whether the ServiceActor::Context is successful. By default, a new
     # context is successful and only changes when explicitly failed.
     #
     # The "success?" method is the inverse of the "failure?" method.
     #
     # Examples
     #
-    #   context = Interactor::Context.new
-    #   # => #<Interactor::Context>
+    #   context = ServiceActor::Context.new
+    #   # => #<ServiceActor::Context>
     #   context.success?
     #   # => true
     #   context.fail!
-    #   # => Interactor::Failure: #<Interactor::Context>
+    #   # => ServiceActor::Failure: #<ServiceActor::Context>
     #   context.success?
     #   # => false
     #
@@ -159,19 +162,19 @@ module Interactor
       !failure?
     end
 
-    # Public: Whether the Interactor::Context has failed. By default, a new
+    # Public: Whether the ServiceActor::Context has failed. By default, a new
     # context is successful and only changes when explicitly failed.
     #
     # The "failure?" method is the inverse of the "success?" method.
     #
     # Examples
     #
-    #   context = Interactor::Context.new
-    #   # => #<Interactor::Context>
+    #   context = ServiceActor::Context.new
+    #   # => #<ServiceActor::Context>
     #   context.failure?
     #   # => false
     #   context.fail!
-    #   # => Interactor::Failure: #<Interactor::Context>
+    #   # => ServiceActor::Failure: #<ServiceActor::Context>
     #   context.failure?
     #   # => true
     #
@@ -180,58 +183,58 @@ module Interactor
       @failure || false
     end
 
-    # Public: Fail the Interactor::Context. Failing a context raises an error
-    # that may be rescued by the calling interactor. The context is also flagged
+    # Public: Fail the ServiceActor::Context. Failing a context raises an error
+    # that may be rescued by the calling actor. The context is also flagged
     # as having failed.
     #
     # Optionally the caller may provide a hash of key/value pairs to be merged
     # into the context before failure.
     #
     # context - A Hash whose key/value pairs are merged into the existing
-    #           Interactor::Context instance. (default: {})
+    #           ServiceActor::Context instance. (default: {})
     #
     # Examples
     #
-    #   context = Interactor::Context.new
-    #   # => #<Interactor::Context>
+    #   context = ServiceActor::Context.new
+    #   # => #<ServiceActor::Context>
     #   context.fail!
-    #   # => Interactor::Failure: #<Interactor::Context>
+    #   # => ServiceActor::Failure: #<ServiceActor::Context>
     #   context.fail! rescue false
     #   # => false
     #   context.fail!(foo: "baz")
-    #   # => Interactor::Failure: #<Interactor::Context foo="baz">
+    #   # => ServiceActor::Failure: #<ServiceActor::Context foo="baz">
     #
-    # Raises Interactor::Failure initialized with the Interactor::Context.
+    # Raises ServiceActor::Failure initialized with the ServiceActor::Context.
     def fail!(context = {})
       context.each { |key, value| modifiable[key.to_sym] = value }
       @failure = true
       raise Failure, self
     end
 
-    # Internal: Track that an Interactor has been called. The "called!" method
-    # is used by the interactor being invoked with this context. After an
-    # interactor is successfully called, the interactor instance is tracked in
+    # Internal: Track that a ServiceActor has been called. The "called!" method
+    # is used by the actor being invoked with this context. After an
+    # actor is successfully called, the actor instance is tracked in
     # the context for the purpose of potential future rollback.
     #
-    # interactor - An Interactor instance that has been successfully called.
+    # actor - A ServiceActor instance that has been successfully called.
     #
     # Returns nothing.
-    def called!(interactor)
-      _called << interactor
+    def called!(actor)
+      _called << actor
     end
 
-    # Public: Roll back the Interactor::Context. Any interactors to which this
+    # Public: Roll back the ServiceActor::Context. Any actors to which this
     # context has been passed and which have been successfully called are asked
     # to roll themselves back by invoking their "rollback" instance methods.
     #
     # Examples
     #
-    #   context = MyInteractor.call(foo: "bar")
-    #   # => #<Interactor::Context foo="baz">
+    #   context = MyActor.call(foo: "bar")
+    #   # => #<ServiceActor::Context foo="baz">
     #   context.rollback!
     #   # => true
     #   context
-    #   # => #<Interactor::Context foo="bar">
+    #   # => #<ServiceActor::Context foo="bar">
     #
     # Returns true if rolled back successfully or false if already rolled back.
     def rollback!
@@ -240,22 +243,22 @@ module Interactor
       @rolled_back = true
     end
 
-    # Internal: An Array of successfully called Interactor instances invoked
-    # against this Interactor::Context instance.
+    # Internal: An Array of successfully called ServiceActor instances invoked
+    # against this ServiceActor::Context instance.
     #
     # Examples
     #
-    #   context = Interactor::Context.new
-    #   # => #<Interactor::Context>
+    #   context = ServiceActor::Context.new
+    #   # => #<ServiceActor::Context>
     #   context._called
     #   # => []
     #
-    #   context = MyInteractor.call(foo: "bar")
-    #   # => #<Interactor::Context foo="baz">
+    #   context = MyActor.call(foo: "bar")
+    #   # => #<ServiceActor::Context foo="baz">
     #   context._called
-    #   # => [#<MyInteractor @context=#<Interactor::Context foo="baz">>]
+    #   # => [#<MyActor @context=#<ServiceActor::Context foo="baz">>]
     #
-    # Returns an Array of Interactor instances or an empty Array.
+    # Returns an Array of ServiceActor instances or an empty Array.
     def _called
       @called ||= []
     end