+++ spec/spec_helper.rb
@@ -1,8 +1,12 @@
-if ENV["CODECLIMATE_REPO_TOKEN"]
-  require "codeclimate-test-reporter"
-  CodeClimate::TestReporter.start
+# frozen_string_literal: true
+
+if ENV["COVERAGE"]
+  require "simplecov"
+  SimpleCov.start do
+    add_filter "/spec/"
+  end
 end
 
-require "interactor"
+require "service_actor"
 
 Dir[File.expand_path("../support/*.rb", __FILE__)].each { |f| require f }